requirements.txt
streamlit
google-genai
gTTS
Pillow